Lots and lots of fixes. The game loads and you can join worlds. A few todos and testing to go over
Regenerated PacketUtils

+++ src/main/java/meteordevelopment/meteorclient/renderer/MeshRenderer.java
@@ -132,7 +132,8 @@ public class MeshRenderer {
         }
 
         int indexCount = mesh != null ? mesh.getIndicesCount()
-            : (indexBuffer != null ? indexBuffer.size() / Integer.BYTES : -1);
+            : (int) (indexBuffer != null ? indexBuffer.size() / Integer.BYTES : -1);
+        // todo hope this is alright @minegame take a look please (lossy conversion from long to int)
 
         if (indexCount > 0) {
